requirements.txt
gradio>=4.0.0
requests
Pillow
transformers>=4.40.0
torch
accelerate
qwen-vl-utils
